+++ etherlink/bin_node/lib_dev/sequencer.ml
@@ -29,7 +29,7 @@ let install_finalizer_seq ~(tx_container : _ Services_backend_sig.tx_container)
     Services_backend_sig.tx_container_module tx_container
   in
   Lwt_exit.register_clean_up_callback ~loc:__LOC__ @@ fun exit_status ->
-  telemetry_cleanup () ;
+  let* () = telemetry_cleanup () in
   let* () = Events.shutdown_node ~exit_status in
   let* () = server_public_finalizer () in
   let* () = server_private_finalizer () in