+++ opentelemetry/src/ambient-context/eio/opentelemetry_ambient_context_eio.mli
@@ -1,2 +1,0 @@
-val storage : unit -> Opentelemetry_ambient_context.storage
-(** Storage using Eio's fibers local storage *)