+++ opentelemetry/src/ambient-context/lwt/opentelemetry_ambient_context_lwt.mli
@@ -1,2 +1,0 @@
-val storage : unit -> Opentelemetry_ambient_context.storage
-(** Storage using Lwt keys *)